Finished basic functions, better performance

+++ src/model/HitBox.java
@@ -8,22 +8,22 @@ public class HitBox {
     private Canvas canvas;
     private GraphicsContext gc;
 
-    private int x;
-    private int y;
+    private double x;
+    private double y;
     private int width;
     private int height;
-    private int centerX;
-    private int centerY;
+    private double centerX;
+    private double centerY;
 
-    public HitBox(Canvas canvas, int x, int y, int width, int height) {
+    public HitBox(Canvas canvas, double x, double y, int width, int height) {
         this.canvas = canvas;
         this.gc = canvas.getGraphicsContext2D();
         this.x = x;
         this.y = y;
         this.width = width;
         this.height = height;
-        centerX = (width / 2) + x;
-        centerY = (height / 2) + y;
+        centerX = ((double) width / 2.0) + x;
+        centerY = ((double) height / 2.0) + y;
     }
 
     public void paint() {
@@ -31,19 +31,19 @@ public class HitBox {
         gc.fillRect(x, y, width, height);
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
 
@@ -63,19 +63,19 @@ public class HitBox {
         this.height = height;
     }
 
-    public int getCenterX() {
+    public double getCenterX() {
         return centerX;
     }
 
-    public void setCenterX(int centerX) {
+    public void setCenterX(double centerX) {
         this.centerX = centerX;
     }
 
-    public int getCenterY() {
+    public double getCenterY() {
         return centerY;
     }
 
-    public void setCenterY(int centerY) {
+    public void setCenterY(double centerY) {
         this.centerY = centerY;
     }
